fixAddClubAdminUserTable
