+++ src/pages/UserTable/ChangeUserClubModal.tsx
@@ -1,14 +1,18 @@
-import React from "react";
+import React, { useState } from "react";
 import { Modal } from "antd";
-import ChangeUserClubForm from "./ChangeUserClubForm";
-import User from "../Distinction/Interfaces/User";
+import NotificationBoxApi from "../../api/NotificationBoxApi";
+import ClubAdmin from "../../models/Club/ClubAdmin";
+import AddAdministratorModal from "../Club/AddAdministratorModal/AddAdministratorModal";
+import ClubUser from "../../models/Club/ClubUser";
+import AdminType from "../../models/Admin/AdminType";
+
 
 interface Props {
   record: string;
   showModal: boolean;
   setShowModal: (showModal: any) => void;
   onChange: (id: string, userRoles: string) => void;
-  user: User
+  user: any
 }
 const ChangeUserClubModal = ({
   record,
@@ -17,22 +21,57 @@ const ChangeUserClubModal = ({
   onChange,
   user
 }: Props) => {
+  const [admin, setAdmin] = useState<ClubAdmin>(new ClubAdmin());
+ 
+  let clubId =user?.clubId;
+  const newAdmin: ClubAdmin = {
+    id: 0,
+    userId: record,
+    user: new ClubUser(),
+    adminType: new AdminType(),
+    clubId: clubId,
+  };
+
+  const handleChange = (id: string, userRole: string) => {
+    onChange(id, userRole);
+    user.clubName &&
+      NotificationBoxApi.createNotifications(
+        [id],
+        `Вам була присвоєна нова роль: '${userRole}' в курені: `,
+        NotificationBoxApi.NotificationTypes.UserNotifications,
+        `/clubs/${clubId}`,user.clubName
+      );
+  };
+
+  const handleClick = async () => {
+    setShowModal(false);
+  };
+  if(!showModal)
+  {
+    return(<div></div>)
+  }
   return (
-    <Modal
-      title="Редагування куреня"
-      visible={showModal}
-      centered
-      footer={null}
-      onCancel={() => setShowModal(false)}
-    >
-      <ChangeUserClubForm
-        record={record}
-        showModal={showModal}
-        setShowModal={setShowModal}
-        onChange={onChange}
-        user={user}
-      />
-    </Modal>
+    <div>
+      {clubId!== null ? (
+        <AddAdministratorModal
+          admin={newAdmin}
+          setAdmin={setAdmin}
+          visibleModal={showModal}
+          setVisibleModal={setShowModal}
+          clubId={clubId}
+          onChange={handleChange}
+        ></AddAdministratorModal>
+        ):
+        (
+          <Modal title="Попередження"
+            visible={showModal}
+            onOk={handleClick}
+            onCancel={handleClick}>
+            <p> Користувач  <b>{user.firstName} {user.lastName} </b>не являється членом курення!{" "}</p>
+            <p>В провід курення можна додати тільки дійсного члена!</p>
+          </Modal>  
+        )};
+    </div>
   );
 };
